Move instrument functions from cli scripts to instrument module
