+++ examples/instrument.py
@@ -1,7 +1,26 @@
+"""Functions for instrumenting and running softlearning examples.
+
+This package contains functions, which allow seamless runs of examples in
+different modes (e.g. locally, in google compute engine, or ec2).
+
+
+There are two types of functions in this file:
+1. run_example_* methods, which run the experiments by invoking
+    `tune.run_experiments` function.
+2. launch_example_* methods, which are helpers function to submit an
+    example to be run in the cloud. In practice, these launch a cluster,
+    and then run the `run_example_cluster` method with the provided
+    arguments and options.
+"""
+
+import importlib
 import os
 import uuid
+from pprint import pformat
 
+import ray
 from ray import tune
+from ray.autoscaler.commands import exec_cluster
 
 from softlearning.misc.utils import datetimestamp, PROJECT_PATH
 
@@ -10,6 +29,80 @@ AUTOSCALER_DEFAULT_CONFIG_FILE_GCE = os.path.join(
     PROJECT_PATH, 'config', 'gcp-ray-autoscaler-mujoco.yaml')
 AUTOSCALER_DEFAULT_CONFIG_FILE_EC2 = os.path.join(
     PROJECT_PATH, 'config', 'aws-ray-autoscaler-mujoco.yaml')
+
+
+def _normalize_trial_resources(resources, cpu, gpu, extra_cpu, extra_gpu):
+    if resources is None:
+        resources = {}
+
+    if cpu is not None:
+        resources['cpu'] = cpu
+
+    if gpu is not None:
+        resources['gpu'] = gpu
+
+    if extra_cpu is not None:
+        resources['extra_cpu'] = extra_cpu
+
+    if extra_gpu is not None:
+        resources['extra_gpu'] = extra_gpu
+
+    return resources
+
+
+def add_command_line_args_to_variant_spec(variant_spec, command_line_args):
+    variant_spec['run_params'].update({
+        'checkpoint_freq': (
+            command_line_args.checkpoint_frequency
+            if command_line_args.checkpoint_frequency is not None
+            else variant_spec['run_params'].get('checkpoint_frequency', 0)
+        ),
+        'checkpoint_at_end': (
+            command_line_args.checkpoint_at_end
+            if command_line_args.checkpoint_at_end is not None
+            else variant_spec['run_params'].get('checkpoint_at_end', True)
+        ),
+    })
+
+    variant_spec['restore'] = command_line_args.restore
+
+    return variant_spec
+
+
+def generate_experiment(trainable_class, variant_spec, command_line_args):
+    local_dir = os.path.join(
+        '~/ray_results',
+        command_line_args.universe,
+        command_line_args.domain,
+        command_line_args.task)
+    resources_per_trial = _normalize_trial_resources(
+        command_line_args.resources_per_trial,
+        command_line_args.trial_cpus,
+        command_line_args.trial_gpus,
+        command_line_args.trial_extra_cpus,
+        command_line_args.trial_extra_gpus)
+
+    datetime_prefix = datetimestamp()
+    experiment_id = '-'.join((datetime_prefix, command_line_args.exp_name))
+
+    variant_spec = add_command_line_args_to_variant_spec(
+        variant_spec, command_line_args)
+
+    experiment = {
+        'run': trainable_class,
+        'resources_per_trial': resources_per_trial,
+        'config': variant_spec,
+        'local_dir': local_dir,
+        'num_samples': command_line_args.num_samples,
+        'upload_dir': command_line_args.upload_dir,
+        'checkpoint_freq': (
+            variant_spec['run_params']['checkpoint_frequency']),
+        'checkpoint_at_end': (
+            variant_spec['run_params']['checkpoint_at_end']),
+        'restore': command_line_args.restore,  # Defaults to None
+    }
+
+    return experiment_id, experiment
 
 
 def unique_cluster_name(args):
@@ -39,3 +132,223 @@ def get_experiments_info(experiments):
     }
 
     return experiments_info
+
+
+def confirm_yes_no(prompt):
+    # raw_input returns the empty string for "enter"
+    yes = {'yes', 'y', 'ye'}
+    no = {'no', 'n'}
+
+    choice = input(prompt).lower()
+    while True:
+        if choice in yes:
+            return True
+        elif choice in no:
+            exit(0)
+        else:
+            print("Please respond with 'yes' or 'no'")
+        choice = input().lower()
+
+
+def run_example_dry(example_module_name, example_argv):
+    """Print the variant spec and related information of an example."""
+    example_module = importlib.import_module(example_module_name)
+
+    example_args = example_module.get_parser().parse_args(example_argv)
+    variant_spec = example_module.get_variant_spec(example_args)
+    trainable_class = example_module.get_trainable_class(example_args)
+
+    experiment_id, experiment = generate_experiment(
+        trainable_class, variant_spec, example_args)
+
+    experiments = {experiment_id: experiment}
+
+    experiments_info = get_experiments_info(experiments)
+    number_of_trials = experiments_info["number_of_trials"]
+    total_number_of_trials = experiments_info["total_number_of_trials"]
+
+    experiments_info_text = f"""
+Dry run.
+
+Experiment specs:
+{pformat(experiments, indent=2)}
+
+Number of trials:
+{pformat(number_of_trials, indent=2)}
+
+Number of total trials (including samples/seeds): {total_number_of_trials}
+"""
+
+    print(experiments_info_text)
+
+
+def run_example_local(example_module_name, example_argv):
+    """Run example locally, potentially parallelizing across cpus/gpus."""
+    example_module = importlib.import_module(example_module_name)
+
+    example_args = example_module.get_parser().parse_args(example_argv)
+    variant_spec = example_module.get_variant_spec(example_args)
+    trainable_class = example_module.get_trainable_class(example_args)
+
+    experiment_id, experiment = generate_experiment(
+        trainable_class, variant_spec, example_args)
+    experiments = {experiment_id: experiment}
+
+    ray.init(
+        num_cpus=example_args.cpus,
+        num_gpus=example_args.gpus,
+        resources=example_args.resources or {},
+        # Tune doesn't currently support local mode
+        local_mode=False,
+        include_webui=example_args.include_webui,
+        temp_dir=example_args.temp_dir)
+
+    tune.run_experiments(
+        experiments,
+        with_server=example_args.with_server,
+        server_port=4321,
+        scheduler=None)
+
+
+def run_example_debug(example_module_name, example_argv):
+    """The debug mode limits tune trial runs to enable use of debugger.
+
+    TODO(hartikainen): The debug mode should allow easy switch between
+    parallelized andnon-parallelized runs such that the debugger can be
+    reasonably used when running the code. This could be implemented for
+    example by requiring a custom resource (e.g. 'debug-resource') that
+    limits the number of parallel runs to one. For this to work, tune needs to
+    merge the support for custom resources:
+    https://github.com/ray-project/ray/pull/2979. Alternatively, this could be
+    implemented using the 'local_mode' argument for ray.init(), once tune
+    supports it.
+    """
+    raise NotImplementedError
+
+
+def run_example_cluster(example_module_name, example_argv):
+    """Run example on cluster mode.
+
+    This functions is very similar to the local mode, except that it
+    correctly sets the redis address to make ray/tune work on a cluster.
+    """
+    example_module = importlib.import_module(example_module_name)
+
+    example_args = example_module.get_parser().parse_args(example_argv)
+    variant_spec = example_module.get_variant_spec(example_args)
+    trainable_class = example_module.get_trainable_class(example_args)
+
+    experiment_id, experiment = generate_experiment(
+        trainable_class, variant_spec, example_args)
+    experiments = {experiment_id: experiment}
+
+    redis_address = ray.services.get_node_ip_address() + ':6379'
+
+    ray.init(
+        redis_address=redis_address,
+        num_cpus=example_args.cpus,
+        num_gpus=example_args.gpus,
+        # Tune doesn't currently support local mode
+        local_mode=False,
+        include_webui=example_args.include_webui,
+        temp_dir=example_args.temp_dir)
+
+    tune.run_experiments(
+        experiments,
+        with_server=example_args.with_server,
+        server_port=4321,
+        scheduler=None)
+
+
+def launch_example_cluster(example_module_name,
+                           example_argv,
+                           config_file,
+                           screen,
+                           tmux,
+                           stop,
+                           start,
+                           override_cluster_name,
+                           port_forward):
+    """Launches the example on autoscaled ray cluster through ray exec_cmd.
+
+    This handles basic validation and sanity checks for the experiment, and
+    then executes the command on autoscaled ray cluster. If necessary, it will
+    also fill in more useful defaults for our workflow (i.e. for tmux and
+    cluster_name).
+    """
+    example_module = importlib.import_module(example_module_name)
+
+    example_args = example_module.get_parser().parse_args(example_argv)
+    variant_spec = example_module.get_variant_spec(example_args)
+    trainable_class = example_module.get_trainable_class(example_args)
+
+    experiment_id, experiment = generate_experiment(
+        trainable_class, variant_spec, example_args)
+    experiments = {experiment_id: experiment}
+
+    experiments_info = get_experiments_info(experiments)
+    total_number_of_trials = experiments_info['total_number_of_trials']
+
+    if not example_args.upload_dir:
+        confirm_yes_no(
+            "`upload_dir` is empty. No results will be uploaded to cloud"
+            " storage. Use `--upload-dir` argument to set upload dir."
+            " Continue without upload directory?\n(yes/no) ")
+
+    confirm_yes_no(f"Launch {total_number_of_trials} trials?\n(yes/no) ")
+
+    override_cluster_name = override_cluster_name or unique_cluster_name(
+        example_args)
+
+    cluster_command_parts = (
+        'softlearning',
+        'run_example_cluster',
+        example_module_name,
+        *example_argv)
+    cluster_command = ' '.join(cluster_command_parts)
+
+    return exec_cluster(
+        config_file=config_file,
+        cmd=cluster_command,
+        screen=screen,
+        tmux=tmux,
+        stop=stop,
+        start=start,
+        override_cluster_name=override_cluster_name,
+        port_forward=port_forward)
+
+
+def launch_example_gce(*args, config_file, **kwargs):
+    """Forwards call to `launch_example_cluster` after adding gce defaults.
+
+    This optionally sets the ray autoscaler configuration file to the default
+    gce configuration file, and then calls `launch_example_cluster` to
+    execute the original command on autoscaled gce cluster by parsing the args.
+
+    See `launch_example_cluster` for further details.
+    """
+    config_file = (
+        config_file or AUTOSCALER_DEFAULT_CONFIG_FILE_GCE)
+
+    return launch_example_cluster(
+        *args,
+        config_file=config_file,
+        **kwargs)
+
+
+def launch_example_ec2(*args, config_file, **kwargs):
+    """Forwards call to `launch_example_cluster` after adding ec2 defaults.
+
+    This optionally sets the ray autoscaler configuration file to the default
+    ec2 configuration file, and then calls `launch_example_cluster` to
+    execute the original command on autoscaled ec2 cluster by parsing the args.
+
+    See `launch_example_cluster` for further details.
+    """
+    config_file = (
+        config_file or AUTOSCALER_DEFAULT_CONFIG_FILE_EC2)
+
+    launch_example_cluster(
+        *args,
+        config_file=config_file,
+        **kwargs)
