+++ examples/utils.py
@@ -285,77 +285,3 @@ def variant_equals(*keys):
         return node
 
     return get_from_spec
-
-
-def _normalize_trial_resources(resources, cpu, gpu, extra_cpu, extra_gpu):
-    if resources is None:
-        resources = {}
-
-    if cpu is not None:
-        resources['cpu'] = cpu
-
-    if gpu is not None:
-        resources['gpu'] = gpu
-
-    if extra_cpu is not None:
-        resources['extra_cpu'] = extra_cpu
-
-    if extra_gpu is not None:
-        resources['extra_gpu'] = extra_gpu
-
-    return resources
-
-
-def add_command_line_args_to_variant_spec(variant_spec, command_line_args):
-    variant_spec['run_params'].update({
-        'checkpoint_freq': (
-            command_line_args.checkpoint_frequency
-            if command_line_args.checkpoint_frequency is not None
-            else variant_spec['run_params'].get('checkpoint_frequency', 0)
-        ),
-        'checkpoint_at_end': (
-            command_line_args.checkpoint_at_end
-            if command_line_args.checkpoint_at_end is not None
-            else variant_spec['run_params'].get('checkpoint_at_end', True)
-        ),
-    })
-
-    variant_spec['restore'] = command_line_args.restore
-
-    return variant_spec
-
-
-def generate_experiment(trainable_class, variant_spec, command_line_args):
-    local_dir = os.path.join(
-        '~/ray_results',
-        command_line_args.universe,
-        command_line_args.domain,
-        command_line_args.task)
-    resources_per_trial = _normalize_trial_resources(
-        command_line_args.resources_per_trial,
-        command_line_args.trial_cpus,
-        command_line_args.trial_gpus,
-        command_line_args.trial_extra_cpus,
-        command_line_args.trial_extra_gpus)
-
-    datetime_prefix = datetimestamp()
-    experiment_id = '-'.join((datetime_prefix, command_line_args.exp_name))
-
-    variant_spec = add_command_line_args_to_variant_spec(
-        variant_spec, command_line_args)
-
-    experiment = {
-        'run': trainable_class,
-        'resources_per_trial': resources_per_trial,
-        'config': variant_spec,
-        'local_dir': local_dir,
-        'num_samples': command_line_args.num_samples,
-        'upload_dir': command_line_args.upload_dir,
-        'checkpoint_freq': (
-            variant_spec['run_params']['checkpoint_frequency']),
-        'checkpoint_at_end': (
-            variant_spec['run_params']['checkpoint_at_end']),
-        'restore': command_line_args.restore,  # Defaults to None
-    }
-
-    return experiment_id, experiment
